# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы электронной очереди",
        "title": "Electronic Queue API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/doctor/active": {
            "get": {
                "description": "Возвращает список всех врачей в системе. Используется для заполнения выпадающих списков на клиенте.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Получить список всех врачей",
                "responses": {
                    "200": {
                        "description": "Массив моделей врачей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Doctor"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/cabinets/active": {
            "get": {
                "description": "Возвращает список всех уникальных номеров кабинетов, когда-либо существовавших в расписании.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Получить список всех существующих кабинетов",
                "responses": {
                    "200": {
                        "description": "Массив номеров кабинетов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/complete-appointment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Завершает прием пациента по талону. Статус талона должен быть 'на_приеме'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Завершить прием пациента",
                "parameters": [
                    {
                        "description": "Данные для завершения приема",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CompleteAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment completed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или статус талона",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/end-break": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Завершает перерыв врача. Статус врача должен быть 'перерыв'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Завершить перерыв врача",
                "parameters": [
                    {
                        "description": "Данные для завершения перерыва",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.EndBreakRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Break ended successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или статус врача",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/queue-all": {
            "get": {
                "description": "Возвращает список всех талонов со статусами 'зарегистрирован' и 'на_приеме' для всех кабинетов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Получить очередь ко всем врачебным кабинетам (для нового табло)",
                "responses": {
                    "200": {
                        "description": "Массив талонов в очереди",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DoctorQueueTicketResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/screen-updates/{cabinet_number}": {
            "get": {
                "description": "Отправляет начальное состояние и последующие обновления статуса приема через Server-Sent Events для конкретного кабинета.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Получить обновления для табло врача",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер кабинета",
                        "name": "cabinet_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Поток событий (см. реальную структуру ответа в коде)",
                        "schema": {
                            "$ref": "#/definitions/handlers.DoctorScreenResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат номера кабинета",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/set-active": {
            "post": {
                "description": "Устанавливает статус врача как активный (при входе в систему).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Установить статус врача как активный",
                "parameters": [
                    {
                        "description": "Данные для установки статуса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SetActiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor status set to active",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/set-inactive": {
            "post": {
                "description": "Устанавливает статус врача как неактивный (при выходе из системы).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Установить статус врача как неактивный",
                "parameters": [
                    {
                        "description": "Данные для установки статуса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SetInactiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor status set to inactive",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/start-appointment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Начинает прием пациента по талону. Статус талона должен быть 'зарегистрирован'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Начать прием пациента",
                "parameters": [
                    {
                        "description": "Данные для начала приема",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.StartAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment started successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или статус талона",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor/start-break": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Начинает перерыв врача. Статус врача должен быть 'активен'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Начать перерыв врача",
                "parameters": [
                    {
                        "description": "Данные для начала перерыва",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.StartBreakRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Break started successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или статус врача",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CompleteAppointmentRequest": {
            "type": "object",
            "required": [
                "ticket_id"
            ],
            "properties": {
                "ticket_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.EndBreakRequest": {
            "type": "object",
            "required": [
                "doctor_id"
            ],
            "properties": {
                "doctor_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.Doctor": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Schedule"
                    }
                },
                "specialization": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.DoctorStatus"
                }
            }
        },
        "models.DoctorQueueTicketResponse": {
            "type": "object",
            "properties": {
                "cabinet_number": {
                    "type": "integer"
                },
                "patient_full_name": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.TicketStatus"
                },
                "ticket_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}